Move enums to their own package
