+++ src/jvmMain/kotlin/GetRecentlyUpdatedItems.jvm.kt
@@ -3,6 +3,8 @@
 package cloud.drakon.ktuniversalis
 
 import cloud.drakon.ktuniversalis.exception.UniversalisException
+import cloud.drakon.ktuniversalis.world.DataCenter
+import cloud.drakon.ktuniversalis.world.World
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.future.future
